% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadXStringSet.R
\name{loadXStringSet}
\alias{loadXStringSet}
\title{Load a DNAStringSet}
\usage{
loadXStringSet(seq.info, project)
}
\arguments{
\item{seq.info}{A named list of metadata for this object.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{DNAStringSet} containing DNA sequences.
This may also be a \linkS4class{QualityScaledDNAStringSet} with quality scores.
}
\description{
Load a \linkS4class{DNAStringSet} object, typically from a FASTA or FASTQ file generated by the corresponding \code{\link{stageObject}} method.
}
\examples{
library(Biostrings)
stuff <- DNAStringSet(c("AAA", "CC", "G", "TTTT"))

tmp <- tempfile()
dir.create(tmp)
info <- stageObject(stuff, tmp, path="dna_thing")
loadXStringSet(info, project=tmp)

}
